+++ java/mod/industrialProcessing/work/recipe/IPowerRecipe.java
@@ -1,5 +1,8 @@
 package mod.industrialProcessing.work.recipe;
 
 public interface IPowerRecipe extends IMachineRecipe {
-    int getPowerRequired();
+	/**
+	 * @return returns the amount of power required per tick
+	 */
+    float getPowerRequired();
 }
